requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Babel==3.1.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-Migrate==4.0.5
WTForms==3.0.1
marshmallow==3.20.1
neo4j>=5.28.2
bcrypt==4.0.1
requests==2.31.0
python-dotenv==1.0.0
python-dateutil==2.8.2
Pillow==12.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0
sentry-sdk==1.32.0
prometheus-client==0.17.1
black==23.9.1
flake8==6.1.0
isort==5.12.0
pytest==7.4.2
pytest-flask==1.2.0
pytest-mock==3.11.1
flasgger==0.9.7.1
reportlab==4.0.4
weasyprint==60.1
PyJWT==2.8.0
email-validator==2.0.0
geopy==2.4.0

# Google Generative AI dependencies
google-generativeai>=0.8.5
google-auth>=2.41.1
google-api-core>=2.25.2
google-api-python-client>=2.184.0
google-auth-oauthlib>=1.2.0